eliminar

+++ app_biblioteca/New Group/Entity/Libro.swift
@@ -12,6 +12,6 @@ struct Libro:Codable
     var id:Int
     var titulo:String
     var autor:Autor
-    var anio_publicacion:Double
+    var anio_publicacion:Int
     var editorial:Editorial
 }
